Rollback uin16_t to int
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/gudhi/trunk@433 636b058d-ea47-450e-bf9e-a15bfbe3eedb

Former-commit-id: 6f25b6c7263c3eca9aa5579a4a5e6eedd9dd0464

+++ src/Persistent_cohomology/include/gudhi/Persistent_cohomology.h
@@ -299,11 +299,11 @@ class Persistent_cohomology {
 
  public:
   /** \brief Initializes the coefficient field.*/
-  void init_coefficients(uint16_t charac) {
+  void init_coefficients(int charac) {
     coeff_field_.init(charac);
   }
   /** \brief Initializes the coefficient field for multi-field persistent homology.*/
-  void init_coefficients(uint16_t charac_min, uint16_t charac_max) {
+  void init_coefficients(int charac_min, int charac_max) {
     coeff_field_.init(charac_min, charac_max);
   }
